Restructure component definitions
